requirements.txt
streamlit==1.39.0
rdkit==2023.9.6
pandas==2.2.3
numpy==2.1.1
torch==2.4.1
torch-geometric==2.5.3
simanneal==0.5.0
pillow==10.4.0
networkx
matplotlib
torchviz
graphvizs